Added segmentation for files
Now the files are divided into sections, where each section begins with the # symbol.  When the chunk with the highest similarity is found for the query the entire file to which that chunk belongs is passed, as a context, to the LLM. Segmentation is performed only for files with .txt and .md extensions.